+++ packages/article-skeleton/fixtures/full-article-content.web.js
@@ -161,6 +161,26 @@ export default [
       {
         attributes: {
           element: {
+            attributes: { "deck-id": "43614" },
+            value: "in-article-info-card"
+          }
+        },
+        children: [],
+        name: "interactive"
+      },
+      {
+        attributes: {
+          element: {
+            attributes: { "deck-id": "43434" },
+            value: "gallery-carousel"
+          }
+        },
+        children: [],
+        name: "interactive"
+      },
+      {
+        attributes: {
+          element: {
             attributes: {
               season: "2020",
               competition: "8",