+++ packages/article-skeleton/src/article-body/article-body.web.js
@@ -29,7 +29,9 @@ import {
   OptaFootballSummary,
   OptaFootballMatchStats,
   OlympicsMedalTable,
-  OlympicsSchedule
+  OlympicsSchedule,
+  InfoCard,
+  GalleryCarousel
 } from "@times-components/ts-components";
 
 import ArticleLink from "./article-link";
@@ -151,6 +153,28 @@ const renderers = ({
   interactive(key, { url, element, display }) {
     const { attributes, value } = element;
     switch (value) {
+      case "in-article-info-card":
+        return (
+          <Context.Consumer key={key}>
+            {({ theme }) => (
+              <FetchProvider url={deckApiUrl + attributes["deck-id"]}>
+                <InfoCard sectionColour={theme.sectionColour} />
+              </FetchProvider>
+            )}
+          </Context.Consumer>
+        );
+
+      case "gallery-carousel":
+        return (
+          <Context.Consumer key={key}>
+            {({ theme }) => (
+              <FetchProvider url={deckApiUrl + attributes["deck-id"]}>
+                <GalleryCarousel sectionColour={theme.sectionColour} />
+              </FetchProvider>
+            )}
+          </Context.Consumer>
+        );
+
       case "newsletter-puff":
         // eslint-disable-next-line no-case-declarations
         const { code, copy, headline, imageUri, label } = attributes;