+++ packages/ts-components/src/index.ts
@@ -47,3 +47,6 @@ export {
 export {
   OlympicsSchedule
 } from './components/olympics/schedule/OlympicsSchedule';
+
+export { InfoCard } from './components/in-article-info-card/InfoCard';
+export { GalleryCarousel } from './components/carousel/GalleryCarousel';