+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.29.30](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.29.29...@times-components/article-magazine-standard@3.29.30) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.29.29](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.29.28...@times-components/article-magazine-standard@3.29.29) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.29.28](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.29.27...@times-components/article-magazine-standard@3.29.28) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.29.27](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.29.26...@times-components/article-magazine-standard@3.29.27) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.29.26](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.29.25...@times-components/article-magazine-standard@3.29.26) (2021-08-02)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 ## [3.29.25](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.29.24...@times-components/article-magazine-standard@3.29.25) (2021-07-29)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard