+++ packages/ts-components/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.25](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.24...@times-components/ts-components@1.11.25) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/ts-components
+
+
+
+
+
+## [1.11.24](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.23...@times-components/ts-components@1.11.24) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/ts-components
+
+
+
+
+
+## [1.11.23](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.22...@times-components/ts-components@1.11.23) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/ts-components
+
+
+
+
+
+## [1.11.22](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.21...@times-components/ts-components@1.11.22) (2021-08-03)
+
+**Note:** Version bump only for package @times-components/ts-components
+
+
+
+
+
+## [1.11.21](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.20...@times-components/ts-components@1.11.21) (2021-08-02)
+
+**Note:** Version bump only for package @times-components/ts-components
+
+
+
+
+
 ## [1.11.20](https://github.com/newsuk/times-components/compare/@times-components/ts-components@1.11.19...@times-components/ts-components@1.11.20) (2021-07-29)